refactored scoreboard into one place
